+++ server/novelgo/api/README.md
@@ -1,0 +1,2 @@
+To generate swagger files:
+swagger generate server -q --server-package ../internal/pkg/restapi --main-package ../../cmd/novelgo-server --model-package ../internal/pkg/models --spec swagger.yml
