--- server/novelgo/cmd/httpserver/README.md
@@ -1,3 +1,0 @@
-To start the server:
-Inside httpserver directory, run the following command
-go run ./cmd/novelgo-server/main.go
