+++ server/novelgo/cmd/novelgo-server/main.go
@@ -9,8 +9,8 @@ import (
 	"github.com/go-openapi/loads"
 	flags "github.com/jessevdk/go-flags"
 
-	"novelgo/cmd/httpserver/restapi"
-	"novelgo/cmd/httpserver/restapi/operations"
+	"novelgo/internal/pkg/restapi"
+	"novelgo/internal/pkg/restapi/operations"
 )
 
 // This file was generated by the swagger tool.
