+++ server/novelgo/internal/pkg/restapi/doc.go
