+++ server/novelgo/internal/pkg/restapi/embedded_spec.go
