+++ server/novelgo/internal/pkg/restapi/operations/create_game_responses.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/go-openapi/runtime"
 
-	"novelgo/cmd/httpserver/models"
+	"novelgo/internal/pkg/models"
 )
 
 // CreateGameCreatedCode is the HTTP code returned for type CreateGameCreated
