+++ server/novelgo/internal/pkg/restapi/operations/create_game_urlbuilder.go
