+++ server/novelgo/internal/pkg/restapi/operations/delete_game_parameters.go
