+++ server/novelgo/internal/pkg/restapi/operations/delete_game_urlbuilder.go
