+++ server/novelgo/internal/pkg/restapi/operations/get_game_by_id.go
