+++ server/novelgo/internal/pkg/restapi/operations/list_games_responses.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/go-openapi/runtime"
 
-	"novelgo/cmd/httpserver/models"
+	"novelgo/internal/pkg/models"
 )
 
 // ListGamesOKCode is the HTTP code returned for type ListGamesOK
