+++ server/novelgo/internal/pkg/restapi/operations/list_games_urlbuilder.go
