+++ server/novelgo/internal/pkg/restapi/operations/novelgo_api.go
