+++ server/novelgo/internal/pkg/restapi/operations/update_game.go
