+++ server/novelgo/internal/pkg/restapi/operations/update_game_urlbuilder.go
